requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pyserial==3.5
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
Pillow==10.4.0
matplotlib==3.9.0
